+++ source/standalone/demo/play_arms.py
@@ -27,18 +27,17 @@ from omni.isaac.orbit.app import AppLauncher
 parser = argparse.ArgumentParser(
     description="This script demonstrates how to use the physics engine to simulate a single-arm manipulator."
 )
-parser.add_argument("--headless", action="store_true", default=False, help="Force display off at all times.")
 parser.add_argument(
     "--robot", type=str, default="franka_panda", choices=["franka_panda", "ur10"], help="Name of the robot."
 )
+# append AppLauncher cli args
+AppLauncher.add_app_launcher_args(parser)
+# parse the arguments
 args_cli = parser.parse_args()
 
 # launch omniverse app
-app_launcher = AppLauncher(headless=args_cli.headless)
+app_launcher = AppLauncher(args_cli)
 simulation_app = app_launcher.app
-
-"""Rest everything follows."""
-
 
 import torch
 import traceback