+++ source/standalone/demo/play_ridgeback_franka.py
@@ -26,12 +26,14 @@ from omni.isaac.orbit.app import AppLauncher
 parser = argparse.ArgumentParser(
     description="This script demonstrates how to simulate a mobile manipulator with dummy joints."
 )
-parser.add_argument("--headless", action="store_true", default=False, help="Force display off at all times.")
 parser.add_argument("--robot", type=str, default="franka_panda", help="Name of the robot.")
+# append AppLauncher cli args
+AppLauncher.add_app_launcher_args(parser)
+# parse the arguments
 args_cli = parser.parse_args()
 
 # launch omniverse app
-app_launcher = AppLauncher(headless=args_cli.headless)
+app_launcher = AppLauncher(args_cli)
 simulation_app = app_launcher.app
 
 """Rest everything follows."""